addressgin pr comments
